Add files via upload
Project Components:

HTML Structure:
The HTML file will provide the basic structure of the web page. It will include the calculator display, buttons for numbers (0-9) and operators (+, -, *, /), as well as the equal (=) button to perform calculations. Additionally, it will have a "Clear" button to reset the calculator.

CSS Styling:
The CSS file will be used to style the calculator, giving it a visually appealing appearance. You can customize the colors, fonts, button styles, and layout to suit your preferences.

JavaScript Functionality:
The JavaScript file will handle the logic and functionality of the calculator. It will include functions for each arithmetic operation (addition, subtraction, multiplication, division), as well as a function to clear the display. When the user clicks on the buttons, the corresponding functions will be executed, and the result will be displayed on the calculator's screen.